requirements.txt
asn1crypto==0.24.0
cffi==1.12.2
cryptography==2.6.1
Django==2.2.4
django-cors-headers==2.5.3
django-extensions==2.1.6
django-filter==2.2.0
django-rest-knox==3.6.0
djangorestframework==3.9.2
gunicorn==19.9.0
Pillow==5.4.1
psycopg2==2.7.7
pycparser==2.19
pyOpenSSL==19.0.0
python-decouple==3.1
pytz==2018.9
six==1.12.0
sqlparse==0.3.0
whitenoise==4.1
